requirements.txt
Flask
tensorflow-cpu
Pillow
numpy
gunicorn
opencv-python-headless
mtcnn
supabase
python-dotenv